requirements.txt
Flask
flask-cors
mysql-connector-python
werkzeug
flask-jwt-extended
gunicorn